requirements.txt
django_graphos==0.3.41
Django==2.1.5
